Change get_var to only accept keys
Prior to this change, get_var would return the value of a variable sent
as an argument. That means passing either a key or a key and value pair would
return the value of the env variable with the same name. Now, the entire
parameter string is considered to be a key.