Add binary and builtin execution
Lots going on. Prior to execution:
- A fork is executed;
- Pipes and redirections are handled;
- For binaries:
    - The command path is resolved;

Builtins are executed twice - once in the fork process (for builtins
that produce output), and once in the parent process. This is necessary,
as output must be affected by redirections, but changes to env must
reflect on the parent process.
For this effect, cd, export and pwd now receive a print_output argument that
controls whether errors and output are written to their respective fds.

Other changes:
- Move redirect_fd() to fd.c;
- Rename resolve_cmd_path() to resolve_binary_path();
- Improve echo() readability;